+++ lib/models/book.dart
@@ -11,6 +11,7 @@ class Book {
   bool isDeleted;
   String? description;
   double rating;
+  int groupId;
   DateTime createTime;
   DateTime updateTime;
 
@@ -25,6 +26,7 @@ class Book {
       required this.isDeleted,
       this.description,
       required this.rating,
+      this.groupId = 0,
       required this.createTime,
       required this.updateTime});
 
@@ -47,8 +49,41 @@ class Book {
       'is_deleted': isDeleted ? 1 : 0,
       'description': description,
       'rating': rating,
+      'group_id': groupId,
       'create_time': createTime.toIso8601String(),
       'update_time': updateTime.toIso8601String(),
     };
+  }
+
+  Book copyWith({
+    int? id,
+    String? title,
+    String? coverPath,
+    String? filePath,
+    String? lastReadPosition,
+    double? readingPercentage,
+    String? author,
+    bool? isDeleted,
+    String? description,
+    double? rating,
+    int? groupId,
+    DateTime? createTime,
+    DateTime? updateTime,
+  }) {
+    return Book(
+      id: id ?? this.id,
+      title: title ?? this.title,
+      coverPath: coverPath ?? this.coverPath,
+      filePath: filePath ?? this.filePath,
+      lastReadPosition: lastReadPosition ?? this.lastReadPosition,
+      readingPercentage: readingPercentage ?? this.readingPercentage,
+      author: author ?? this.author,
+      isDeleted: isDeleted ?? this.isDeleted,
+      description: description ?? this.description,
+      rating: rating ?? this.rating,
+      groupId: groupId ?? this.groupId,
+      createTime: createTime ?? this.createTime,
+      updateTime: updateTime ?? this.updateTime,
+    );
   }
 }